I added The Boss Rooms, and Starting Room
Created The Boss Rooms for both the Depths, and the Reef (The Reef Boss Room has two versions for whether or not we are able to implement Poison Sea. The Version with poison Sea uses circles to indicate where the Poison Sea WONT be).

Did not create starting room for the reef, as this can simply use the current starting room created by the art team.